requirements.txt
bcrypt==3.1.6
certifi==2018.11.29
cffi==1.11.5
Django==1.11.17
Pillow==5.4.1
pycparser==2.19
pytz==2018.9
six==1.12.0
wincertstore==0.2
